+++ package.json
@@ -10,6 +10,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@notion-render/client": "^0.0.2",
     "@notionhq/client": "^3.1.3",
     "@types/stopword": "^2.0.3",
     "@types/tiny-segmenter": "^0.2.0",
