+++ src/app/globals.css
@@ -1,4 +1,5 @@
 @import "tailwindcss";
+@import "./notion.css";
 
 :root {
   --background: var(--color-n-bg);
